fix warning

+++ Horion/Command/Commands/CoordsCommand.cpp
@@ -12,6 +12,6 @@ CoordsCommand::~CoordsCommand()
 bool CoordsCommand::execute(std::vector<std::string>* args)
 {
 	vec3_t* pos = g_Data.getLocalPlayer()->getPos();
-	clientMessageF("[%sHorion%s] %sX: %.1f Y: %.1f Z: %.1f", GOLD, WHITE, GREEN, pos->x, pos->y - 1.62f /* eye height */, pos->z);
+	clientMessageF("[%sHorion%s] %sX: %.1f Y: %.1f Z: %.1f", GOLD, WHITE, GREEN, pos->x, (float) (pos->y - 1.62f) /* eye height */, pos->z);
 	return true;
 }
